requirements.txt
Django>=5.0,<6.0
django-ckeditor>=6.7.0,<7.0
Pillow>=10.0,<11.0
python-dotenv>=1.0,<2.0
PyMySQL>=1.1,<2.0
